requirements.txt
tensorflow
pandas 
PyYAML 
numpy
matplotlib
seaborn
opencv-python
pillow
ipykernel